Major refactor: Use canvas drawing to display ranch editor
