+++ package.json
@@ -18,15 +18,13 @@
     "next": "15.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-tooltip": "^5.28.0",
-    "react-virtualized-auto-sizer": "^1.0.25",
-    "react-window": "^1.8.11"
+    "react-tooltip": "^5.28.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@types/file-saver": "^2.0.7",
     "@types/node": "^20",
-    "@types/react": "^19",
+    "@types/react": "^19.0.2",
     "@types/react-dom": "^19",
     "@types/react-window": "^1.8.8",
     "eslint": "^9",
